+++ src/main/java/de/fraunhofer/isst/configmanager/connector/dataspaceconnector/clients/DataspaceBrokerClient.java
@@ -15,7 +15,6 @@ import org.springframework.stereotype.Service;
 
 import java.io.IOException;
 import java.net.URI;
-import java.util.Objects;
 import java.util.UUID;
 
 @Slf4j